+++ src/main/java/com/alisa/moviereservationsystem/dto/paymentDto/PaymentRequestDto.java
@@ -1,4 +1,6 @@
 package com.alisa.moviereservationsystem.dto.paymentDto;
 
-public record PaymentRequestDto(Long reservationId, Long userId, Float generalPrice) {
+import java.math.BigDecimal;
+
+public record PaymentRequestDto(Long reservationId, Long userId, BigDecimal generalPrice) {
 }
