+++ src/main/java/com/alisa/moviereservationsystem/dto/updateDto/SeatUpdateDto.java
@@ -2,6 +2,8 @@ package com.alisa.moviereservationsystem.dto.updateDto;
 
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record SeatUpdateDto(
         @NotNull(message = "Seat number must not be null")
         Integer seatNumber,
@@ -10,5 +12,5 @@ public record SeatUpdateDto(
         Integer rowNumber,
 
         @NotNull(message = "Price must not be null")
-        Float price) {
+        BigDecimal price) {
 }
