+++ src/main/java/com/alisa/moviereservationsystem/models/Reservation.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -25,7 +26,7 @@ public class Reservation {
     @CreationTimestamp
     private Instant timeStamp;
 
-    private Float generalPrice;
+    private BigDecimal generalPrice;
 
     @Enumerated(EnumType.STRING)
     private ReservationStatus status;
